Moved files to lib folder
