+++ lib/resources.ts
