+++ package.json
@@ -17,7 +17,7 @@
     "dataloader": "^2.0.0",
     "graphql": "^15.4.0",
     "graphql-tag": "^2.11.0",
-    "handlebars": "^4.7.6",
+    "handlebars": "^4.7.7",
     "lodash.groupby": "^4.6.0",
     "lodash.isequal": "^4.5.0",
     "lodash.mapvalues": "^4.6.0",
